producers/news/requirements.txt
kafka-python==2.0.2
feedparser==6.0.11
python-dateutil==2.9.0
requests


